hash config into instance

+++ packages/protocol/contracts/L1/TaikoData.sol
@@ -14,9 +14,9 @@ library TaikoData {
         // This number is calculated from maxNumProposedBlocks to make
         // the 'the maximum value of the multiplier' close to 20.0
         uint256 maxVerificationsPerTx;
-        uint256 blockMaxGasLimit;
-        uint256 maxTransactionsPerBlock;
-        uint256 maxBytesPerTxList;
+        uint64 blockMaxGasLimit;
+        uint64 maxTransactionsPerBlock;
+        uint64 maxBytesPerTxList;
         uint256 txListCacheExpiry;
         uint256 proofCooldownPeriod;
         uint256 systemProofCooldownPeriod;
